% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compare_column_sets}
\alias{compare_column_sets}
\title{Compare values between two data frame columns}
\usage{
compare_column_sets(df1, col1, df2, col2)
}
\arguments{
\item{df1}{First data frame.}

\item{col1}{Column in `df1` (unquoted).}

\item{df2}{Second data frame.}

\item{col2}{Column in `df2` (unquoted).}
}
\value{
A list with elements `only_in_df1` and `only_in_df2`.
}
\description{
Compares two columns from two data frames and returns the values
that are only present in one but not the other.
}
\examples{
 \dontrun{
df1 <- tibble::tibble(id = c(1, 2, 3))
df2 <- tibble::tibble(id = c(3, 4, 5))
compare_column_sets(df1, id, df2, id)
}

}
