% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpmObjects.R
\docType{data}
\name{facility_size_order}
\alias{facility_size_order}
\title{Ordered facility size levels}
\format{
A character vector of length 7.
}
\usage{
facility_size_order
}
\description{
A character vector defining the standard order of facility types by size.
Useful for setting factor levels for consistent sorting in plots and tables.
}
\examples{
factor(c("CHC", "SHC", "RH"), levels = facility_size_order, ordered = TRUE)

}
\keyword{datasets}
