% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hw_assessment}
\alias{hw_assessment}
\title{Health Workforce Service Standard Cutoffs}
\format{

}
\usage{
hw_assessment
}
\description{
This dataset contains minimum service standard cutoff values used to assess health facilities by type.
Values are defined separately for Sub-Health Centers (SHC) and Primary/Referral Hospitals (PH/RH).
}
\examples{
hw_assessment
}
\keyword{datasets}
