% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tpm_data.R
\name{print.tpm_tools_list}
\alias{print.tpm_tools_list}
\title{Print method for tpm_tools_list objects}
\usage{
\method{print}{tpm_tools_list}(x, ...)
}
\arguments{
\item{x}{An object of class `tpm_tools_list` (as returned by [list_tpm_tools()]).}

\item{...}{Additional arguments passed to [print()].}
}
\description{
Nicely formats the grouped data, gt tables, colors, and functions in the tpmTools package.
}
