% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{scale_color_tpm}
\alias{scale_color_tpm}
\alias{scale_fill_tpm}
\title{ggplot2 Scales using TPM Palettes}
\usage{
scale_color_tpm(palette = "cat", ...)

scale_fill_tpm(palette = "cat", ...)
}
\arguments{
\item{palette}{Character name of the palette to return.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_*_manual}.}
}
\description{
Color and fill scales for ggplot2 based on TPM official palettes.
}
\examples{
 \dontrun{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt, color = factor(cyl))) +
  geom_point(size = 3) +
  scale_color_tpm("cat")

ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_tpm("seq_blue")
}
}
