% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{data}
\name{tpm_colors}
\alias{tpm_colors}
\alias{tpm_blue_dark}
\alias{tpm_blue_light}
\alias{tpm_green_light}
\alias{tpm_green_dark}
\alias{tpm_green_mid}
\alias{tpm_red}
\alias{tpm_yellow}
\alias{tpm_maroon}
\alias{tpm_purple}
\alias{tpm_gray_dark}
\alias{tpm_gray_mid}
\alias{tpm_gray_light}
\title{TPM Official Colors}
\format{
Character scalars of HEX codes.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 12.
}
\usage{
tpm_blue_dark

tpm_blue_light

tpm_green_light

tpm_green_dark

tpm_green_mid

tpm_red

tpm_yellow

tpm_maroon

tpm_purple

tpm_gray_dark

tpm_gray_mid

tpm_gray_light

tpm_colors
}
\description{
Individual hex values for the official TPM Monitoring color system,
aligned with glitr-style palettes and designed for use in data visualization.
These colors are optimized for accessibility and colorblind-friendly contrasts.
}
\examples{
\dontrun{
tpm_red
tpm_blue_dark
tpm_colors["red"]
names(tpm_colors)
}
}
\keyword{datasets}
