% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{tpm_show_palettes}
\alias{tpm_show_palettes}
\title{Show TPM Palettes}
\usage{
tpm_show_palettes(
  palettes = c("cat", "seq_blue", "seq_red", "div_blue_red", "div_green_purple"),
  ncol = 1
)
}
\arguments{
\item{palettes}{Character vector of palette names to display.
Defaults to all available palettes.}

\item{ncol}{Number of columns in the display.}
}
\description{
Display TPM official palettes as swatches for quick reference.
This function is inspired by \code{glitr::si_palettes}.
}
\examples{
\donttest{
tpm_show_palettes()
tpm_show_palettes(c("cat", "seq_blue"))
}
}
